requirements.txt
asgiref==3.8.1
Django==5.0.6
djangorestframework==3.15.2
joblib==1.4.2
numpy==1.23.5
scikit-learn==1.5.0
scipy==1.13.1
sqlparse==0.5.0
threadpoolctl==3.5.0
tzdata==2024.1
xgboost==2.0.3
pandas==1.5.3

